+++ src/ic-cdk/src/api/management_canister/main/mod.rs
@@ -6,45 +6,31 @@
 //! [1]: https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-management-canister
 
 use crate::api::call::{call, call_with_payment128, CallResult};
+use crate::api::canister_version;
 use candid::Principal;
 
 mod types;
 pub use types::*;
 
-/// Cycles cost to create a canister.
-///
-/// See [Computation and Storage Costs](https://internetcomputer.org/docs/current/developer-docs/deploy/computation-and-storage-costs)
-pub const CREATE_CANISTER_CYCLES: u128 = 100_000_000_000u128;
-
 /// Register a new canister and get its canister id.
 ///
-/// Note: This call charges [CREATE_CANISTER_CYCLES] from the caller canister.
-///
 /// See [IC method `create_canister`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-create_canister).
-pub async fn create_canister(arg: CreateCanisterArgument) -> CallResult<(CanisterIdRecord,)> {
-    call_with_payment128(
-        Principal::management_canister(),
-        "create_canister",
-        (arg,),
-        CREATE_CANISTER_CYCLES,
-    )
-    .await
-}
-
-/// [create_canister] and specify extra cycles to the new canister.
 ///
-/// Note: This call charges [CREATE_CANISTER_CYCLES] and the specified extra cycles from the caller canister.
-///
-/// See [IC method `create_canister`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-create_canister).
-pub async fn create_canister_with_extra_cycles(
+/// This call requires cycles payment. The required cycles varies according to the subnet size (number of nodes).
+/// Check [Gas and cycles cost](https://internetcomputer.org/docs/current/developer-docs/gas-cost) for more details.
+pub async fn create_canister(
     arg: CreateCanisterArgument,
     cycles: u128,
 ) -> CallResult<(CanisterIdRecord,)> {
+    let extended_arg = CreateCanisterArgumentExtended {
+        settings: arg.settings,
+        sender_canister_version: Some(canister_version()),
+    };
     call_with_payment128(
         Principal::management_canister(),
         "create_canister",
-        (arg,),
-        CREATE_CANISTER_CYCLES + cycles,
+        (extended_arg,),
+        cycles,
     )
     .await
 }
@@ -53,47 +39,78 @@ pub async fn create_canister_with_extra_cycles(
 ///
 /// See [IC method `update_settings`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-update_settings).
 pub async fn update_settings(arg: UpdateSettingsArgument) -> CallResult<()> {
-    call(Principal::management_canister(), "update_settings", (arg,)).await
+    let extended_arg = UpdateSettingsArgumentExtended {
+        canister_id: arg.canister_id,
+        settings: arg.settings,
+        sender_canister_version: Some(canister_version()),
+    };
+    call(
+        Principal::management_canister(),
+        "update_settings",
+        (extended_arg,),
+    )
+    .await
 }
 
 /// Install code into a canister.
 ///
 /// See [IC method `install_code`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-install_code).
 pub async fn install_code(arg: InstallCodeArgument) -> CallResult<()> {
-    call(Principal::management_canister(), "install_code", (arg,)).await
+    let extended_arg = InstallCodeArgumentExtended {
+        mode: arg.mode,
+        canister_id: arg.canister_id,
+        wasm_module: arg.wasm_module,
+        arg: arg.arg,
+        sender_canister_version: Some(canister_version()),
+    };
+    call(
+        Principal::management_canister(),
+        "install_code",
+        (extended_arg,),
+    )
+    .await
 }
 
 /// Remove a canister's code and state, making the canister empty again.
 ///
-/// See [IC method `uninstall_code`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-uninstall_code)
+/// See [IC method `uninstall_code`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-uninstall_code).
 pub async fn uninstall_code(arg: CanisterIdRecord) -> CallResult<()> {
-    call(Principal::management_canister(), "uninstall_code", (arg,)).await
+    let extended_arg = CanisterIdRecordExtended {
+        canister_id: arg.canister_id,
+        sender_canister_version: Some(canister_version()),
+    };
+    call(
+        Principal::management_canister(),
+        "uninstall_code",
+        (extended_arg,),
+    )
+    .await
 }
 
 /// Start a canister if the canister status was `stopped` or `stopping`.
 ///
-/// See [IC method `start_canister`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-start_canister)
+/// See [IC method `start_canister`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-start_canister).
 pub async fn start_canister(arg: CanisterIdRecord) -> CallResult<()> {
     call(Principal::management_canister(), "start_canister", (arg,)).await
 }
 
 /// Stop a canister.
 ///
-/// See [IC method `stop_canister`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-stop_canister)
+/// See [IC method `stop_canister`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-stop_canister).
 pub async fn stop_canister(arg: CanisterIdRecord) -> CallResult<()> {
     call(Principal::management_canister(), "stop_canister", (arg,)).await
 }
 
 /// Get status information about the canister.
 ///
-/// See [IC method `canister_status`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-canister_status)
+/// See [IC method `canister_status`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-canister_status).
 pub async fn canister_status(arg: CanisterIdRecord) -> CallResult<(CanisterStatusResponse,)> {
     call(Principal::management_canister(), "canister_status", (arg,)).await
 }
 
 /// Delete a canister from the IC.
 ///
-/// See [IC method `delete_canister`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-delete_canister)
+/// See [IC method `delete_canister`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-delete_canister).
 pub async fn delete_canister(arg: CanisterIdRecord) -> CallResult<()> {
     call(Principal::management_canister(), "delete_canister", (arg,)).await
 }
@@ -103,7 +120,7 @@ pub async fn delete_canister(arg: CanisterIdRecord) -> CallResult<()> {
 /// Note that, beyond the argument as specified in the interface description,
 /// there is a second parameter `cycles` which is the amount of cycles to be deposited.
 ///
-/// See [IC method `deposit_cycles`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-deposit_cycles)
+/// See [IC method `deposit_cycles`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-deposit_cycles).
 pub async fn deposit_cycles(arg: CanisterIdRecord, cycles: u128) -> CallResult<()> {
     call_with_payment128(
         Principal::management_canister(),
@@ -116,7 +133,14 @@ pub async fn deposit_cycles(arg: CanisterIdRecord, cycles: u128) -> CallResult<(
 
 /// Get 32 pseudo-random bytes.
 ///
-/// See [IC method `raw_rand`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-raw_rand)
+/// See [IC method `raw_rand`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-raw_rand).
 pub async fn raw_rand() -> CallResult<(Vec<u8>,)> {
     call(Principal::management_canister(), "raw_rand", ()).await
+}
+
+/// Get public information about the canister.
+///
+/// See [IC method `canister_info`](https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-canister_info).
+pub async fn canister_info(arg: CanisterInfoRequest) -> CallResult<(CanisterInfoResponse,)> {
+    call(Principal::management_canister(), "canister_info", (arg,)).await
 }