+++ library/ic-ledger-types/CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.7.0] - 2023-07-13
+
+### Added
+- from_hex/from_slice/to_hex methods to AccountIdentifier in ic-ledger-types
+
+### Changed
+- Upgrade `ic-cdk` to v0.10 and `candid` to v0.9.
+
+## [0.6.0] - 2023-06-20
+### Changed
+- Upgrade `ic-cdk` to v0.9.
+
 ## [0.5.0] - 2023-05-26
 ### Changed
 - Upgrade `ic-cdk` to v0.8.