+++ src/ic-cdk-macros/CHANGELOG.md
@@ -6,6 +6,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- Export composite_query to Candid. (#419)
+
+## [0.7.1] - 2023-07-27
+
+### Fixed
+
+- Only update/query macros can take guard function. (#417)
+
+## [0.7.0] - 2023-07-13
+
+### Added
+
+- `export_candid` macro. (#386)
+
+### Changed
+
+- Remove `import` macro. (#390)
+
 ## [0.6.10] - 2023-03-01
 
 ### Changed